+++ playwright.config.js
@@ -9,9 +9,6 @@ if (process.argv.includes('--project=composer')) {
 } else if (process.argv.includes('--project=core')) {
   process.env.port = '7201';
   options          = '-t ./test/core/cases';
-} else if (process.argv.includes('--project=core.viewport')) {
-  process.env.port = '7211';
-  options          = '-t ./test/core.viewport/cases';
 } else if (process.argv.includes('--project=svg')) {
   process.env.port = '7202';
   options          = '-i /src/lib/gsvg.script.js ' +
@@ -57,6 +54,12 @@ if (process.argv.includes('--project=composer')) {
                      '-i /src/plugins/template.engine.script.js ' +
                      '-i /src/plugins/load.script.js ' +
                      '-t test/plugins/load/cases';
+} else if (process.argv.includes('--project=core.intersection')) {
+  process.env.port = '7211';
+  options          = '-t ./test/core.intersection/cases';
+} else if (process.argv.includes('--project=plugin.intersection')) {
+  process.env.port = '7212';
+  options          = '-t ./test/plugin.intersection/cases';
 }
 
 const webServer = process.env.port ? {
@@ -70,7 +73,6 @@ export default defineConfig({
   projects      : [
     {name : 'helpers'},
     {name : 'core'},
-    {name : 'core.viewport'},
     {name : 'composer'},
     {name : 'svg'},
     {name : 'animateto'},
@@ -81,6 +83,8 @@ export default defineConfig({
     {name : 'template.engine'},
     {name : 'shapes'},
     {name : 'load'},
+    {name : 'core.intersection'},
+    {name : 'plugin.intersection'},
   ],
   testDir       : './test',
   fullyParallel : true,