+++ src/plugins/template.engine.script.js
@@ -1,4 +1,11 @@
-import {svgPlugin} from './template.engine.js';
+import svgPlugin from './template.engine.js';
+
 if (gSVG) {
   gSVG.install(svgPlugin);
 }
+
+customElements
+  .whenDefined('g-composer')
+  .then(composer => {
+    composer.install(svgPlugin);
+  });